+++ src/main/java/dev/willyelton/crystal_tools/CrystalTools.java
@@ -4,6 +4,7 @@ import dev.willyelton.crystal_tools.gui.ModGUIs;
 import dev.willyelton.crystal_tools.levelable.block.ModBlocks;
 import dev.willyelton.crystal_tools.command.RegisterCommandEvent;
 import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
+import dev.willyelton.crystal_tools.crafting.ItemDisabledCondition;
 import dev.willyelton.crystal_tools.crafting.ModRecipes;
 import dev.willyelton.crystal_tools.levelable.ModItems;
 import dev.willyelton.crystal_tools.levelable.armor.ModArmor;
@@ -11,6 +12,7 @@ import dev.willyelton.crystal_tools.network.PacketHandler;
 import dev.willyelton.crystal_tools.levelable.tool.ModTools;
 import dev.willyelton.crystal_tools.world.ModOres;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.common.crafting.CraftingHelper;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
@@ -27,6 +29,7 @@ public class CrystalTools {
     // Directly reference a log4j logger.
     private static final Logger LOGGER = LogManager.getLogger();
     public static final String MODID = "crystal_tools";
+    public static final ItemDisabledCondition.Serializer INSTANCE = new ItemDisabledCondition.Serializer();
 
     public CrystalTools() {
         IEventBus eventBus = FMLJavaModLoadingContext.get().getModEventBus();
@@ -74,6 +77,9 @@ public class CrystalTools {
 
     private void init(final FMLCommonSetupEvent event) {
         // some preinit code
+        event.enqueueWork(() -> {
+            CraftingHelper.register(INSTANCE);
+        });
     }
 
     private void setup() {