+++ src/main/java/dev/willyelton/crystal_tools/gui/UpgradeScreen.java
@@ -1,13 +1,15 @@
 package dev.willyelton.crystal_tools.gui;
 
+import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.gui.component.SkillButton;
+import dev.willyelton.crystal_tools.levelable.ModItems;
 import dev.willyelton.crystal_tools.levelable.skill.SkillDataNode;
-import dev.willyelton.crystal_tools.network.PacketHandler;
-import dev.willyelton.crystal_tools.network.ToolAttributePacket;
-import dev.willyelton.crystal_tools.network.ToolHealPacket;
+import dev.willyelton.crystal_tools.network.*;
 import dev.willyelton.crystal_tools.levelable.LevelableItem;
 import dev.willyelton.crystal_tools.levelable.skill.SkillData;
+import dev.willyelton.crystal_tools.utils.InventoryUtils;
 import dev.willyelton.crystal_tools.utils.NBTUtils;
+import dev.willyelton.crystal_tools.utils.ToolUtils;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
@@ -15,8 +17,11 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 
+import java.util.Set;
+
 public class UpgradeScreen extends BaseUpgradeScreen {
     private Button healButton;
+    private Button resetButton;
     private final ItemStack stack;
     final CompoundTag tag;
 
@@ -52,6 +57,37 @@ public class UpgradeScreen extends BaseUpgradeScreen {
             Component text = Component.literal("Uses a skill point to fully repair this tool");
             UpgradeScreen.this.renderTooltip(poseStack, UpgradeScreen.this.minecraft.font.split(text, Math.max(UpgradeScreen.this.width / 2 - 43, 170)), mouseX, mouseY);
         }));
+
+        resetButton = addRenderableWidget(new Button(width - 40 - 5, 15, 40, Y_SIZE, Component.literal("Reset"), (button) -> {
+            boolean requiresCrystal = CrystalToolsConfig.REQUIRE_CRYSTAL_FOR_RESET.get();
+
+            if (!requiresCrystal || this.player.getInventory().hasAnyOf(Set.of(ModItems.CRYSTAL.get()))) {
+                // Server
+                PacketHandler.sendToServer(new ResetSkillsPacket());
+                PacketHandler.sendToServer(new RemoveItemPacket(ModItems.CRYSTAL.get().getDefaultInstance()));
+
+                // Client
+                ToolUtils.resetPoints(this.stack);
+                InventoryUtils.removeItemFromInventory(this.player.getInventory(), ModItems.CRYSTAL.get().getDefaultInstance());
+
+                int[] points = NBTUtils.getIntArray(stack, "points");
+                if (stack.getItem() instanceof LevelableItem) {
+                    String toolType = ((LevelableItem) stack.getItem()).getItemType();
+                    data = SkillData.fromResourceLocation(new ResourceLocation("crystal_tools", String.format("skill_trees/%s.json", toolType)), points);
+                } else {
+                    data = null;
+                }
+            }
+
+            this.onClose();
+        }, (button, poseStack, mouseX, mouseY) -> {
+            boolean requiresCrystal = CrystalToolsConfig.REQUIRE_CRYSTAL_FOR_RESET.get();
+            String text = "Reset Skill Points";
+            if (requiresCrystal) text += " (Requires 1 Crystal)";
+            Component textComponent = Component.literal(text);
+            UpgradeScreen.this.renderTooltip(poseStack, UpgradeScreen.this.minecraft.font.split(textComponent, Math.max(UpgradeScreen.this.width / 2 - 43, 170)), mouseX, mouseY);
+        }));
+
     }
 
     @Override