+++ src/main/java/dev/willyelton/crystal_tools/levelable/ModItems.java
@@ -2,6 +2,7 @@ package dev.willyelton.crystal_tools.levelable;
 
 import dev.willyelton.crystal_tools.CreativeTabs;
 import dev.willyelton.crystal_tools.CrystalTools;
+import dev.willyelton.crystal_tools.levelable.tool.CrystalApple;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.registries.DeferredRegister;
@@ -15,6 +16,7 @@ public class ModItems {
     // Items
     public static final RegistryObject<Item> CRYSTAL = ITEMS.register("crystal", () -> new Item(new Item.Properties().tab(CreativeTabs.CRYSTAL_TOOLS_TAB)));
     public static final RegistryObject<Item> NETHERITE_STICK = ITEMS.register("netherite_stick", () -> new Item(new Item.Properties().fireResistant().tab(CreativeTabs.CRYSTAL_TOOLS_TAB)));
+    public static final RegistryObject<Item> CRYSTAL_APPLE = ITEMS.register("crystal_apple", CrystalApple::new);
 
     public static void initItems() {
         ITEMS.register(FMLJavaModLoadingContext.get().getModEventBus());