+++ src/main/java/dev/willyelton/crystal_tools/levelable/tool/AIOLevelableTool.java
@@ -2,6 +2,7 @@ package dev.willyelton.crystal_tools.levelable.tool;
 
 import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Multimap;
+import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.keybinding.KeyBindings;
 import dev.willyelton.crystal_tools.utils.NBTUtils;
 import dev.willyelton.crystal_tools.utils.StringUtils;
@@ -33,12 +34,17 @@ public class AIOLevelableTool extends DiggerLevelableTool {
 
     @Override
     public boolean correctTool(ItemStack tool, BlockState blockState) {
-        return true;
+        return blockState.getDestroySpeed(null, null) != -1;
     }
 
     // From Sword
     @Override
     public boolean hurtEnemy(ItemStack tool, @NotNull LivingEntity target, @NotNull LivingEntity attacker) {
+        if (this.isDisabled()) {
+            tool.shrink(1);
+            return false;
+        }
+
         tool.hurtAndBreak(1, attacker, (p_43296_) -> {
             p_43296_.broadcastBreakEvent(EquipmentSlot.MAINHAND);
         });
@@ -78,6 +84,11 @@ public class AIOLevelableTool extends DiggerLevelableTool {
     @Override
     public @NotNull InteractionResult useOn(UseOnContext context) {
         ItemStack stack = context.getItemInHand();
+
+        if (this.isDisabled()) {
+            stack.shrink(1);
+            return InteractionResult.FAIL;
+        }
 
         UseMode mode = UseMode.fromString(NBTUtils.getString(stack, "use_mode"));
 
@@ -128,5 +139,10 @@ public class AIOLevelableTool extends DiggerLevelableTool {
 
     public InteractionResult useOnHoe(UseOnContext context) {
         return ToolUseUtils.useOnHoe(context, this);
+    }
+
+    @Override
+    public boolean isDisabled() {
+        return CrystalToolsConfig.DISABLE_AIOT.get();
     }
 }