+++ src/main/java/dev/willyelton/crystal_tools/levelable/tool/AxeLevelableTool.java
@@ -1,5 +1,6 @@
 package dev.willyelton.crystal_tools.levelable.tool;
 
+import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.keybinding.KeyBindings;
 import dev.willyelton.crystal_tools.utils.NBTUtils;
 import dev.willyelton.crystal_tools.utils.ToolUseUtils;
@@ -31,6 +32,10 @@ public class AxeLevelableTool extends LevelableTool {
 
     @Override
     public @NotNull InteractionResult useOn(@NotNull UseOnContext pContext) {
+        if (this.isDisabled()) {
+            pContext.getItemInHand().shrink(1);
+            return InteractionResult.FAIL;
+        }
         InteractionResult result = ToolUseUtils.useOnAxe(pContext, this);
 
         if (result == InteractionResult.SUCCESS || result == InteractionResult.sidedSuccess(pContext.getLevel().isClientSide)) {
@@ -94,13 +99,10 @@ public class AxeLevelableTool extends LevelableTool {
             recursiveBreakHelper(tool, level, blockPos, entity, minedBlock, 0);
         }
 
-        super.mineBlock(tool, level, blockState, blockPos, entity);
-
-        return true;
+        return super.mineBlock(tool, level, blockState, blockPos, entity);
     }
 
     private void recursiveBreakHelper(ItemStack tool, Level level, BlockPos blockPos, LivingEntity entity, Block block, int depth) {
-
         if (depth > MAX_RECURSIVE_DEPTH) {
             return;
         }
@@ -130,5 +132,10 @@ public class AxeLevelableTool extends LevelableTool {
     @Override
     public boolean correctTool(ItemStack tool, BlockState blockState) {
         return super.correctTool(tool, blockState) || (NBTUtils.getFloatOrAddKey(tool, "leaf_mine") > 0 && blockState.is(BlockTags.MINEABLE_WITH_HOE));
+    }
+
+    @Override
+    public boolean isDisabled() {
+        return CrystalToolsConfig.DISABLE_AXE.get();
     }
 }