+++ src/main/java/dev/willyelton/crystal_tools/levelable/tool/LevelableTool.java
@@ -12,6 +12,7 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.tags.TagKey;
 import net.minecraft.util.Mth;
 import net.minecraft.world.Container;
+import net.minecraft.world.InteractionResult;
 import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
@@ -31,7 +32,7 @@ import java.util.Optional;
 import java.util.function.Consumer;
 
 // For now just focus on things that mine (not sword)
-public class LevelableTool extends Item implements LevelableItem {
+public abstract class LevelableTool extends Item implements LevelableItem {
 
     // Blocks that can be mined by default, null for none
     protected final TagKey<Block> blocks;
@@ -66,12 +67,23 @@ public class LevelableTool extends Item implements LevelableItem {
 
     @Override
     public boolean hurtEnemy(ItemStack tool, @NotNull LivingEntity target, @NotNull LivingEntity attacker) {
+        if (this.isDisabled()) {
+            tool.shrink(1);
+            return false;
+        }
+
         tool.hurtAndBreak(2, attacker, (player) -> player.broadcastBreakEvent(EquipmentSlot.MAINHAND));
         return true;
     }
 
     @Override
     public boolean mineBlock(@NotNull ItemStack tool, Level level, @NotNull BlockState blockState, @NotNull BlockPos blockPos, @NotNull LivingEntity entity) {
+        // If this tool is disabled break on use
+        if (this.isDisabled()) {
+            tool.shrink(1);
+            return false;
+        }
+
         if (!level.isClientSide && blockState.getDestroySpeed(level, blockPos) != 0.0F) {
             tool.hurtAndBreak(1, entity, (player) -> player.broadcastBreakEvent(EquipmentSlot.MAINHAND));