+++ src/main/java/dev/willyelton/crystal_tools/levelable/tool/ShovelLevelableTool.java
@@ -1,5 +1,6 @@
 package dev.willyelton.crystal_tools.levelable.tool;
 
+import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.utils.ToolUseUtils;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.world.InteractionResult;
@@ -16,11 +17,20 @@ public class ShovelLevelableTool extends DiggerLevelableTool {
 
     @Override
     public @NotNull InteractionResult useOn(@NotNull UseOnContext pContext) {
+        if (this.isDisabled()) {
+            pContext.getItemInHand().shrink(1);
+            return InteractionResult.FAIL;
+        }
         return ToolUseUtils.useOnShovel(pContext, this);
     }
 
     @Override
     public boolean canPerformAction(ItemStack stack, net.minecraftforge.common.ToolAction toolAction) {
         return net.minecraftforge.common.ToolActions.DEFAULT_SHOVEL_ACTIONS.contains(toolAction);
+    }
+
+    @Override
+    public boolean isDisabled() {
+        return CrystalToolsConfig.DISABLE_SHOVEL.get();
     }
 }