+++ src/main/java/dev/willyelton/crystal_tools/utils/ArrayUtils.java
@@ -7,13 +7,22 @@ public class ArrayUtils {
                 return true;
             }
         }
-
         return false;
+    }
+
+    public static <T> int indexOf(T[] array, T element) {
+        for (int i = 0; i < array.length; i++) {
+            if (element.equals(array[i])) {
+                return i;
+            }
+        }
+
+        return -1;
     }
 
     public static int indexOf(int[] array, int element) {
         for (int i = 0; i < array.length; i++) {
-            if (array[i] == element) {
+            if (element == array[i]) {
                 return i;
             }
         }